+++ src/Plugins/SmartStore.AmazonPay/Controllers/AmazonPayShoppingCartController.cs
@@ -12,10 +12,9 @@ namespace SmartStore.AmazonPay.Controllers
 			_apiService = apiService;
 		}
 
-		public ActionResult LoginHandler(string orderReferenceId)
+		public ActionResult PayButtonHandler()
 		{
-			var model = _apiService.ProcessPluginRequest(AmazonPayRequestType.LoginHandler, TempData, orderReferenceId);
-
+			var model = _apiService.CreateViewModel(AmazonPayRequestType.PayButtonHandler, TempData);
 			return GetActionResult(model);
 		}
 
@@ -24,7 +23,7 @@ namespace SmartStore.AmazonPay.Controllers
 		{
 			if (ControllerContext.ParentActionViewContext.RequestContext.RouteData.IsRouteEqual("ShoppingCart", "Cart"))
 			{
-				var model = _apiService.ProcessPluginRequest(AmazonPayRequestType.ShoppingCart, TempData);
+				var model = _apiService.CreateViewModel(AmazonPayRequestType.ShoppingCart, TempData);
 
 				return GetActionResult(model);
 			}
@@ -36,7 +35,7 @@ namespace SmartStore.AmazonPay.Controllers
 		{
 			if (renderAmazonPayView)
 			{
-				var model = _apiService.ProcessPluginRequest(AmazonPayRequestType.OrderReviewData, TempData);
+				var model = _apiService.CreateViewModel(AmazonPayRequestType.OrderReviewData, TempData);
 
 				return View(model);
 			}
@@ -48,24 +47,25 @@ namespace SmartStore.AmazonPay.Controllers
 		{
 			if (renderAmazonPayView)
 			{
-				var model = _apiService.ProcessPluginRequest(AmazonPayRequestType.MiniShoppingCart, TempData);
+				var model = _apiService.CreateViewModel(AmazonPayRequestType.MiniShoppingCart, TempData);
 
 				return GetActionResult(model);
 			}
 			return new EmptyResult();
 		}
 
-		[ChildActionOnly]
-		public ActionResult WidgetLibrary()
-		{
-			// not possible to load it asynchronously cause of document.write inside
-			string widgetUrl = _apiService.GetWidgetUrl();
+		//[ChildActionOnly]
+		//public ActionResult WidgetLibrary()
+		//{
+		//	// Not possible to load it asynchronously cause of document.write inside.
+		//	var widgetUrl = _apiService.GetWidgetUrl();
 
-			if (widgetUrl.HasValue())
-			{
-				return this.Content("<script src=\"{0}\" type=\"text/javascript\"></script>".FormatWith(widgetUrl));
-			}
-			return new EmptyResult();
-		}
+		//	if (widgetUrl.HasValue())
+		//	{
+		//		return this.Content("<script src=\"{0}\" type=\"text/javascript\"></script>".FormatWith(widgetUrl));
+		//	}
+
+		//	return new EmptyResult();
+		//}
 	}
 }