+++ src/Plugins/SmartStore.AmazonPay/Events/MessageTokenEventConsumer.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using SmartStore.AmazonPay.Services;
 using SmartStore.Core.Domain.Messages;
 using SmartStore.Core.Events;
 using SmartStore.Core.Plugins;
@@ -7,7 +6,6 @@ using SmartStore.Services;
 using SmartStore.Services.Messages;
 using SmartStore.Services.Orders;
 using SmartStore.Web.Framework;
-using System;
 
 namespace SmartStore.AmazonPay.Events
 {
@@ -34,13 +32,13 @@ namespace SmartStore.AmazonPay.Events
 
 			var storeId = _services.StoreContext.CurrentStore.Id;
 
-			if (!_pluginFinder.IsPluginReady(_services.Settings, AmazonPayCore.SystemName, storeId))
+			if (!_pluginFinder.IsPluginReady(_services.Settings, AmazonPayPlugin.SystemName, storeId))
 				return;
 
             var orderId = messageTokenEvent.Tokens.Where(x => x.Key.Equals("Order.ID")).FirstOrDefault();
-            var order = _orderService.GetOrderById(Convert.ToInt32(orderId.Value));
+            var order = _orderService.GetOrderById(orderId.Value.ToInt());
 
-            var isAmazonPayment = (order != null && order.PaymentMethodSystemName.IsCaseInsensitiveEqual(AmazonPayCore.SystemName));
+			var isAmazonPayment = (order != null && order.PaymentMethodSystemName.IsCaseInsensitiveEqual(AmazonPayPlugin.SystemName));
 			var tokenValue = (isAmazonPayment ? _services.Localization.GetResource("Plugins.Payments.AmazonPay.BillingAddressMessageNote") : "");
 
 			messageTokenEvent.Tokens.Add(new Token("SmartStore.AmazonPay.BillingAddressMessageNote", tokenValue));