+++ src/Plugins/SmartStore.AmazonPay/Models/AmazonPayViewModel.cs
@@ -1,6 +1,6 @@
-﻿using System;
-using SmartStore.AmazonPay.Services;
+﻿using SmartStore.AmazonPay.Services;
 using SmartStore.Web.Framework.Modelling;
+using SmartStore.Web.Models.Common;
 
 namespace SmartStore.AmazonPay.Models
 {
@@ -12,48 +12,42 @@ namespace SmartStore.AmazonPay.Models
 			RedirectAction = "Cart";
 			RedirectController = "ShoppingCart";
 			Result = AmazonPayResultType.PluginView;
-			WidgetUrl = AmazonPayCore.UrlWidgetProduction.FormatWith("de");
+			BillingAddress = new AddressModel();
 		}
 
-		public string SystemName { get { return AmazonPayCore.SystemName; } }
+		public string SystemName
+		{
+			get { return AmazonPayPlugin.SystemName; }
+		}
 
 		public string SellerId { get; set; }
 		public string ClientId { get; set; }
 
+		/// <summary>
+		/// Amazon widget script URL
+		/// </summary>
 		public string WidgetUrl { get; set; }
-		public string ButtonUrl { get; set; }
-		public string LoginHandlerUrl { get; set; }
+		public string ButtonHandlerUrl { get; set; }
 
 		public bool IsShippable { get; set; }
 		public bool IsRecurring { get; set; }
 
+		public string LanguageCode { get; set; }
 		public AmazonPayRequestType Type { get; set; }
 		public AmazonPayResultType Result { get; set; }
+		
 		public string RedirectAction { get; set; }
 		public string RedirectController { get; set; }
 
 		public string OrderReferenceId { get; set; }
+		public string AddressConsentToken { get; set; }
 		public string Warning { get; set; }
 
-		public int AddressWidgetWidth { get; set; }
-		public int AddressWidgetHeight { get; set; }
-
-		public int PaymentWidgetWidth { get; set; }
-		public int PaymentWidgetHeight { get; set; }
-
-		public bool DisplayRewardPoints { get; set; }
-		public int RewardPointsBalance { get; set; }
-		public string RewardPointsAmount { get; set; }
-		public bool UseRewardPoints { get; set; }
+		public string ButtonType { get; set; }
+		public string ButtonColor { get; set; }
+		public string ButtonSize { get; set; }
 
 		public string ShippingMethod { get; set; }
-
-		public string GetWidgetId
-		{
-			get
-			{
-				return "AmazonPay" + Type.ToString();
-			}
-		}
-    }
+		public AddressModel BillingAddress { get; set; }
+	}
 }